requirements.txt
anyio==4.11.0
blinker==1.9.0
certifi==2025.11.12
click==8.3.1
colorama==0.4.6
dnspython==2.8.0
exceptiongroup==1.3.0
Flask==3.1.2
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
pyaes==1.6.1
pymongo==4.15.4
Pyrogram==2.0.106
PySocks==1.7.1
python-telegram-bot==22.5
rapidfuzz==3.5.2
sniffio==1.3.1
TgCrypto==1.2.5
typing_extensions==4.15.0
Werkzeug==3.1.3
